feat(xtask): hold every file to a production-line limit
`cargo xtask structure` caps a source file at **800 lines of production code**,
and runs in pre-commit and CI. It takes about 0.07s: it reads files and counts.

## It counts production lines, not total lines

This workspace keeps most tests inline and gates a hard 100%, so roughly two
thirds of the tree is test code. A total-lines rule would fire hardest on the
best-tested files, which is exactly backwards. The count stops at the first
column-zero `#[cfg(test)]` - indented ones sit on a field or a method and do not
begin the test section - and sibling test files are skipped outright, matching
cargo-llvm-cov's own exclusion so the two cannot drift apart.

## It is a ratchet, not a flat cap

800 is the workspace's 95th percentile (894 today, against a median of 218), so
23 files already sit above it. Setting the cap where the tree does not yet reach
produces a build that cannot go green, which in practice becomes an exemption
list that only grows.

So `KNOWN_LONG` records those 23 with the size they had today, rounded up to the
next 50 lines so an unrelated one-line edit does not fail the build. They may
shrink freely and may not grow; everything else is held to 800 outright.

The check also fails on a **stale** row - a budget whose file is gone, or now
under the cap. That is the failure mode a ratchet actually has: the file gets
split, nobody deletes the row, and the next long file to take that path inherits
an exemption nobody granted it.

## What it would have caught

`blueprint.rs` at 1,571, `config.rs` at 1,576 and `host.rs` at 2,063 all grew
past any reasonable limit without anything objecting. Each is now split; the
rule is what stops the next one.

xtask is inside the walk, so the tooling is held to the same limit as the code -
`structure.rs` itself is 250 lines.

The seam is a reader function, so the whole rule is tested without a filesystem;
one test does walk the real tree, because a structural check that never runs
against the actual workspace is decorative. 122 xtask tests pass.
