+++ crates/leviath-cli/src/daemon/spawn/mod.rs
@@ -35,795 +35,18 @@ use crate::daemon::tool_service::{AgentToolState, CliToolService};
 /// Default max sub-agent tree depth when a blueprint doesn't set one.
 const DEFAULT_SUBAGENT_DEPTH: usize = 3;
 
-/// The user's default provider/model from `config.toml`, in the plain form the
-/// runtime's stage resolver takes.
-pub(crate) fn model_defaults(config: &Config) -> ModelDefaults {
-    ModelDefaults {
-        provider: config.default_provider.clone(),
-        model: config.default_model.clone(),
-        fallback_order: parse_fallback_order(&config.providers.fallback_order),
-    }
-}
-
-/// Parse `[providers] fallback_order` entries (`"provider/model"`) into the
-/// runtime's own form.
-///
-/// A malformed entry is dropped with a warning rather than failing the load: a
-/// typo in a *safety net* should not stop the daemon from starting, and the
-/// warning says which entry went nowhere. Splitting on the first `/` keeps
-/// model ids that contain one (`deepseek/deepseek-v4-flash`) intact.
-/// The `[security] shell_env` decision for this daemon, in the shape the tools
-/// layer wants.
-///
-/// One resolver rather than three: the built-in `shell` tool, a Rhai `shell()`,
-/// and a region's command seed all hand the daemon's environment to a child, so
-/// they answer to the same setting. A script that has `shell` would otherwise be
-/// the way around the `env_var` gate.
-fn shell_env_policy(config: &Config) -> leviath_tools::ShellEnvPolicy {
-    leviath_tools::ShellEnvPolicy {
-        mode: config.security.shell_env,
-        allow_env_vars: config.security.allow_env_vars.clone(),
-        withhold: config.security.shell_env_withhold.clone(),
-    }
-}
-
-fn parse_fallback_order(entries: &[String]) -> Vec<leviath_core::blueprint::ModelEntry> {
-    entries
-        .iter()
-        .filter_map(|raw| match raw.split_once('/') {
-            Some((provider, model)) if !provider.is_empty() && !model.is_empty() => Some(
-                leviath_core::blueprint::ModelEntry::new(provider.to_string(), model.to_string()),
-            ),
-            _ => {
-                tracing::warn!(
-                    entry = %raw,
-                    "ignoring [providers] fallback_order entry: expected \"provider/model\""
-                );
-                None
-            }
-        })
-        .collect()
-}
-
-/// The directories scanned for an agent's Rhai script tools, in precedence order
-/// (earlier wins on a name collision): the agent's own `<agent_dir>/tools/`, then
-/// `extra` (the run workdir's `tools/`, only for `dynamic_tools` agents so a
-/// mid-run write is picked up), then the global `~/.leviath/tools/`. `Option`'s
-/// iterator flattens the "no parent" / "no home" cases without a dangling
-/// `if let` else region.
-fn script_scan_dirs(
-    blueprint_path: &str,
-    extra: Option<std::path::PathBuf>,
-) -> Vec<std::path::PathBuf> {
-    std::path::Path::new(blueprint_path)
-        .parent()
-        .map(|d| d.join("tools"))
-        .into_iter()
-        .chain(extra)
-        .chain(leviath_core::tools_dir())
-        .collect()
-}
-
-/// Read and compile every custom region's Rhai script declared by `blueprint`
-/// (global layout plus each stage's per-stage layout), keyed by the script
-/// path as written. Paths resolve relative to the blueprint's directory (the
-/// script-tool convention - the script travels with the agent), with absolute
-/// paths passing through `Path::join` unchanged. Each distinct path is read
-/// and compiled once; regions sharing a script share the compiled AST.
-///
-/// A missing or uncompilable script is a **hard spawn error** (fail fast,
-/// before any tokens are spent): a hook that silently never ran would change
-/// every inference with no signal. Runtime hook *eval* failures, by contrast,
-/// warn and fall back per hook.
-/// Compile every output validator the blueprint names, keyed by path.
-///
-/// A hard spawn error for the same reason a region script is: the moment to
-/// discover an agent cannot check its own answer is not the end of a long run,
-/// which is the only other time this script would ever be read.
-///
-/// Paths resolve against the blueprint directory, the same convention script
-/// tools and region hooks use, so a validator travels with the agent that needs
-/// it.
-/// Resolve a blueprint-declared script path against the blueprint's directory,
-/// refusing anything that lands outside it.
-///
-/// A script is code the blueprint ships, so it has no business living anywhere
-/// but beside the blueprint. Without this check `base.join(declared)` happily
-/// accepts `../../../../etc/shadow`: the file is read, handed to the Rhai
-/// compiler, and whether it compiles becomes an oracle for what exists on the
-/// host - from a package `lev add` installed, which `SECURITY.md` treats as a
-/// real attacker. There is no `[read_paths]` fallback here on purpose: that
-/// mechanism exists for an agent reading *data* the user pointed it at, and no
-/// legitimate agent loads its own logic from outside its own directory.
-fn script_within_blueprint(
-    base: &std::path::Path,
-    declared: &str,
-    what: &str,
-) -> Result<std::path::PathBuf, String> {
-    let full = base.join(declared);
-    match leviath_core::resolves_within(&full, base) {
-        true => Ok(full),
-        false => Err(format!(
-            "{what} '{declared}' resolves outside the blueprint's directory ({}); a script must \
-             live beside the agent that declares it",
-            base.display()
-        )),
-    }
-}
-
-pub(crate) fn resolve_output_validators(
-    blueprint: &Blueprint,
-    blueprint_path: &str,
-) -> Result<HashMap<String, Arc<leviath_scripting::output_validator::OutputValidator>>, String> {
-    let base = std::path::Path::new(blueprint_path)
-        .parent()
-        .map(std::path::Path::to_path_buf)
-        .unwrap_or_default();
-    let mut compiled = HashMap::new();
-
-    let specs = blueprint
-        .output
-        .iter()
-        .chain(blueprint.stages.iter().filter_map(|s| s.output.as_ref()));
-    for spec in specs {
-        let Some(script) = spec.validator.as_deref() else {
-            continue;
-        };
-        if compiled.contains_key(script) {
-            continue;
-        }
-        let path = script_within_blueprint(&base, script, "output validator")?;
-        let source = std::fs::read_to_string(&path)
-            .map_err(|e| format!("cannot read output validator '{}': {e}", path.display()))?;
-        let validator = leviath_scripting::output_validator::compile(script, &source)
-            .map_err(|e| format!("output validator failed to compile: {e}"))?;
-        compiled.insert(script.to_string(), Arc::new(validator));
-    }
-    Ok(compiled)
-}
-
-/// Compile every stage-hook script the blueprint declares, keyed by the path as
-/// written (issue #260).
-///
-/// Fail-fast at spawn, exactly as region scripts are: an unreadable file, one
-/// that does not compile, or one the blueprint names for a hook it does not
-/// define is a spawn error rather than a surprise partway through a run. One
-/// file backing several hooks is read and compiled once.
-///
-/// Returns an empty map when no stage declares a hook, so the agent gets a
-/// component that every lookup misses rather than a special case.
-pub(crate) fn resolve_stage_hook_scripts(
-    blueprint: &Blueprint,
-    blueprint_path: &str,
-) -> Result<HashMap<String, Arc<leviath_scripting::stage_hook::HookScript>>, String> {
-    let base = std::path::Path::new(blueprint_path)
-        .parent()
-        .map(std::path::Path::to_path_buf)
-        .unwrap_or_default();
-
-    // Gather what each file is wanted for before compiling, so a file backing
-    // two hooks is checked for both in one pass.
-    let mut wanted: HashMap<&str, Vec<&str>> = HashMap::new();
-    for stage in &blueprint.stages {
-        for (hook, path) in stage.hooks.declared() {
-            wanted.entry(path).or_default().push(hook);
-        }
-    }
-
-    let mut scripts = HashMap::new();
-    for (path, hooks) in wanted {
-        let full = script_within_blueprint(&base, path, "stage hook script")?;
-        let source = std::fs::read_to_string(&full)
-            .map_err(|e| format!("cannot read stage hook script '{}': {e}", full.display()))?;
-        let compiled = leviath_scripting::stage_hook::compile(path, &source, &hooks)
-            .map_err(|e| format!("stage hook script '{path}' failed to compile: {e}"))?;
-        scripts.insert(path.to_string(), Arc::new(compiled));
-    }
-    Ok(scripts)
-}
-
-pub(crate) fn resolve_region_scripts(
-    blueprint: &Blueprint,
-    blueprint_path: &str,
-) -> Result<HashMap<String, Arc<leviath_scripting::region_hook::RegionScript>>, String> {
-    let base = std::path::Path::new(blueprint_path)
-        .parent()
-        .map(std::path::Path::to_path_buf)
-        .unwrap_or_default();
-    let mut scripts = HashMap::new();
-
-    let layouts = std::iter::once(&blueprint.context_layout).chain(
-        blueprint
-            .stages
-            .iter()
-            .filter_map(|s| s.context_layout.as_ref()),
-    );
-    for layout in layouts {
-        for region in &layout.regions {
-            let leviath_core::RegionKind::Custom { script, .. } = &region.kind else {
-                continue;
-            };
-            if scripts.contains_key(script) {
-                continue;
-            }
-            let path = script_within_blueprint(&base, script, "custom region script")
-                .map_err(|e| format!("region '{}': {e}", region.name))?;
-            let source = std::fs::read_to_string(&path).map_err(|e| {
-                format!(
-                    "region '{}': cannot read custom region script '{}': {e}",
-                    region.name,
-                    path.display()
-                )
-            })?;
-            let compiled =
-                leviath_scripting::region_hook::compile(script, &source).map_err(|e| {
-                    format!(
-                        "region '{}': custom region script failed to compile: {e}",
-                        region.name
-                    )
-                })?;
-            scripts.insert(script.clone(), Arc::new(compiled));
-        }
-    }
-    Ok(scripts)
-}
-
-/// Names already claimed by a built-in, sub-agent, or MCP tool - a discovered
-/// script tool colliding with one of these is dropped (never shadows a core tool).
-fn reserved_tool_names(builtin_names: &HashSet<String>, mcp_tool_defs: &[Tool]) -> HashSet<String> {
-    let mut reserved: HashSet<String> = builtin_names.clone();
-    reserved.extend(leviath_tools::BuiltinTools::subagent_tool_names());
-    reserved.extend(mcp_tool_defs.iter().map(|t| t.name.clone()));
-    reserved
-}
-
-/// Map a script's self-declared `@requires` capability name to the platform
-/// [`ToolCapability`] it corresponds to. An unrecognized name returns `None`,
-/// which the discovery pass treats as unsatisfiable (the tool is dropped) so a
-/// typo can't silently slip a tool through the platform gate.
-fn script_cap(name: &str) -> Option<leviath_tools::ToolCapability> {
-    match name {
-        "network" | "net" | "http" => Some(leviath_tools::ToolCapability::Network),
-        "shell" | "process" | "process_spawn" => Some(leviath_tools::ToolCapability::ProcessSpawn),
-        "filesystem" | "file" | "fs" => Some(leviath_tools::ToolCapability::FileSystem),
-        _ => None,
-    }
-}
-
-/// Whether `platform` can satisfy every capability a script `@requires`. An
-/// unknown capability name is never satisfiable.
-fn platform_satisfies_caps(
-    platform: &leviath_tools::PlatformCapabilities,
-    required_caps: &[String],
-) -> bool {
-    required_caps
-        .iter()
-        .all(|c| script_cap(c).is_some_and(|cap| platform.supports(cap)))
-}
-
-/// Whether the *current* platform can satisfy a script's `@requires` - the same
-/// gate `discover_script_tools_in` applies at spawn. Exposed so the read-only CLI
-/// surfaces (`lev tools`, `lev validate`, `lev mcp list`) report a tool's real
-/// availability (and flag an unknown/typo'd capability) instead of listing a tool
-/// the daemon would silently drop.
-pub(crate) fn current_platform_satisfies(required_caps: &[String]) -> bool {
-    platform_satisfies_caps(
-        &leviath_tools::PlatformCapabilities::current(),
-        required_caps,
-    )
-}
-
-/// Discover and compile the script tools in `dirs`, returning the compiled set,
-/// the routable names (collisions against `reserved` excluded), and the
-/// advertised `Tool` defs.
-///
-/// A tool whose `@requires` capabilities the current platform can't satisfy is
-/// dropped here (self-declared platform gating) - mirroring how
-/// built-ins filter against [`PlatformCapabilities`].
-pub(crate) fn discover_script_tools_in(
-    dirs: &[std::path::PathBuf],
-    reserved: &HashSet<String>,
-) -> (leviath_scripting::ScriptToolSet, HashSet<String>, Vec<Tool>) {
-    let (set, skipped) = leviath_scripting::ScriptToolSet::discover(dirs);
-    for s in &skipped {
-        // Pre-format the path to a plain string so the `tracing` field carries no
-        // inline method call (an inline `%s.path.display()` leaves a macro
-        // sub-region llvm-cov can't attribute even with the event enabled).
-        let path = s.path.display().to_string();
-        tracing::warn!(tool = %path, reason = %s.reason, "skipping invalid script tool");
-    }
-    let platform = leviath_tools::PlatformCapabilities::current();
-    let mut names = HashSet::new();
-    let mut defs = Vec::new();
-    for meta in set.metas() {
-        if reserved.contains(&meta.name) {
-            tracing::warn!(tool = %meta.name, "script tool name collides with an existing tool - ignoring");
-            continue;
-        }
-        if !platform_satisfies_caps(&platform, &meta.required_caps) {
-            let caps = meta.required_caps.join(", ");
-            tracing::warn!(tool = %meta.name, requires = %caps, "script tool requires a capability this platform lacks - ignoring");
-            continue;
-        }
-        names.insert(meta.name.clone());
-        defs.push(Tool {
-            name: meta.name.clone(),
-            description: meta.description.clone(),
-            parameters: meta.parameters_schema(),
-        });
-    }
-    (set, names, defs)
-}
-
-/// Discover the agent's Rhai script tools and build their `Tool`
-/// defs (the spawn-time entry point). `extra_dir` adds the run workdir's `tools/`
-/// for `dynamic_tools` agents.
-fn discover_script_tools(
-    blueprint_path: &str,
-    builtin_names: &HashSet<String>,
-    mcp_tool_defs: &[Tool],
-    extra_dir: Option<std::path::PathBuf>,
-) -> (leviath_scripting::ScriptToolSet, HashSet<String>, Vec<Tool>) {
-    let dirs = script_scan_dirs(blueprint_path, extra_dir);
-    let reserved = reserved_tool_names(builtin_names, mcp_tool_defs);
-    discover_script_tools_in(&dirs, &reserved)
-}
-
-/// Build one agent's [`AgentToolState`] from the shared executors + config.
-///
-/// `stage_perms_by_index` holds every stage's `[tool_permissions]` (in stage
-/// order); the entry stage's map seeds `stage_perms`, and the pipeline's
-/// `sync_stage` swaps in the right one as the agent changes stage.
-/// Everything [`build_tool_state`] assembles an agent's tool state from.
-///
-/// A struct rather than twenty-one positional parameters, and the reason is
-/// narrower than the lint: three of them - `run_id`, `entry_stage` and
-/// `agent_name` - are all `&str`. Transposing any two of those compiles
-/// silently and produces a run whose approvals are keyed to the wrong name.
-/// Nothing else in this file was ever going to catch that.
-struct ToolStateParts<'a> {
-    /// The built-in tools, over this agent's workdir.
-    builtins: Arc<leviath_tools::BuiltinTools>,
-    /// Their names, for deciding what is a builtin at dispatch.
-    builtin_names: HashSet<String>,
-    /// MCP connections shared across agents.
-    mcp: Arc<Mutex<leviath_mcp::ToolExecutor>>,
-    /// The resolved daemon configuration.
-    config: &'a Config,
-    /// Where this agent's prompts are parked.
-    hub: &'a InteractionHub,
-    /// This run's id, which grants are recorded against.
-    run_id: &'a str,
-    /// The stage the agent enters at.
-    entry_stage: &'a str,
-    /// That stage's index in the blueprint.
-    entry_index: usize,
-    /// Per-stage tool policies, indexed by stage.
-    stage_perms_by_index: Vec<HashMap<String, String>>,
-    /// Per-stage required tools, indexed by stage.
-    stage_required_by_index: Vec<HashSet<String>>,
-    /// Agent-wide tool policies from the blueprint.
-    agent_perms: HashMap<String, String>,
-    /// The blueprint's name, for policy lookup and messages.
-    agent_name: &'a str,
-    /// `--allow` / `--yolo` overrides for this launch.
-    launch_overrides: HashMap<String, crate::config::ToolPolicy>,
-    /// Handle for the sub-agent tools, when this agent may spawn.
-    subagent: Option<SubAgentHandle>,
-    /// The sandbox shell calls run in, when one is configured.
-    sandbox: Option<Arc<crate::daemon::sandbox_manager::SandboxManager>>,
-    /// Rhai tools discovered for this agent.
-    script_tools: leviath_scripting::ScriptToolSet,
-    /// Their names, kept apart so a rescan can diff against them.
-    script_tool_names: HashSet<String>,
-    /// The host those scripts call back into.
-    script_host: Arc<dyn leviath_scripting::ScriptHost>,
-    /// Re-resolution context, for a blueprint that rescans mid-run.
-    dynamic: Option<Arc<crate::daemon::tool_service::DynamicToolCtx>>,
-    /// Whether this run answers its own prompts (`--yolo`).
-    unattended: bool,
-    /// `[safe_commands]` the blueprint declares, if the user opted in.
-    blueprint_safe: Option<&'a leviath_core::blueprint::SafeCommandsConfig>,
-}
-
-fn build_tool_state(parts: ToolStateParts<'_>) -> Arc<AgentToolState> {
-    let entry_perms = parts
-        .stage_perms_by_index
-        .get(parts.entry_index)
-        .cloned()
-        .unwrap_or_default();
-    let entry_required = parts
-        .stage_required_by_index
-        .get(parts.entry_index)
-        .cloned()
-        .unwrap_or_default();
-    Arc::new(AgentToolState {
-        // One budget per run, so the per-run ceiling spans every batch rather
-        // than resetting with each one.
-        writes: Arc::new(crate::daemon::tool_service::WriteBudget::new(
-            parts.config.limits.write_limits(),
-        )),
-        builtins: parts.builtins,
-        mcp: parts.mcp,
-        builtin_names: parts.builtin_names,
-        launch_overrides: Arc::new(parts.launch_overrides),
-        safe_keys: Arc::new(
-            parts
-                .config
-                .safe_keys_for_agent(parts.agent_name, parts.blueprint_safe)
-                .into_keys()
-                .collect(),
-        ),
-        run_allows: Arc::new(Mutex::new(HashSet::new())),
-        stage_allows: Arc::new(StdMutex::new(HashSet::new())),
-        stage_allows_index: Arc::new(StdMutex::new(None)),
-        stage_perms: Arc::new(StdMutex::new(entry_perms)),
-        stage_perms_by_index: Arc::new(parts.stage_perms_by_index),
-        stage_required: Arc::new(StdMutex::new(entry_required)),
-        stage_required_by_index: Arc::new(parts.stage_required_by_index),
-        agent_perms: Arc::new(parts.agent_perms),
-        blueprint_may_loosen: parts.config.security.allow_blueprint_permissions,
-        // The ceiling a blueprint may tighten but not loosen: the user's global
-        // `[tool_permissions]` plus any `[agent_tool_permissions.<name>]` grant
-        // they made for this specific agent. Resolved once here so every later
-        // `resolve_policy` reads one flat map.
-        global_perms: Arc::new(parts.config.permissions_for_agent(parts.agent_name)),
-        interaction: parts.hub.backend_for(parts.run_id),
-        unattended: parts.unattended,
-        stage_name: Arc::new(StdMutex::new(parts.entry_stage.to_string())),
-        subagent: parts.subagent,
-        sandbox: parts.sandbox,
-        script_tools: Arc::new(StdMutex::new(parts.script_tools)),
-        script_tool_names: Arc::new(StdMutex::new(parts.script_tool_names)),
-        script_host: parts.script_host,
-        dynamic: parts.dynamic,
-    })
-}
-
-/// Resolve a blueprint-declared seed path against the run's working directory.
-///
-/// The same rule the `read_file` tool follows, and for the same reason: the
-/// *blueprint* chose this path, not the user, so an installed package could
-/// otherwise write `seed = { files = ["../../.leviath/config.toml"] }` and have
-/// the provider keys in that file seeded straight into a pinned context region -
-/// from where they travel to the model, and out through the answer, a webhook,
-/// or a sub-agent. `read_file` has been confined for exactly this since the
-/// symlink-escape fix; seeding was the path that stayed open.
-///
-/// Outside the workdir falls back to `[read_paths]`, which is already the
-/// mechanism for "this agent is meant to read there and the user agreed",
-/// rather than a second answer to the same question.
-fn seed_path_within(
-    base: &std::path::Path,
-    declared: &std::path::Path,
-    read_paths: &leviath_core::ReadPathPolicy,
-) -> Result<std::path::PathBuf, String> {
-    if leviath_core::resolves_within(declared, base) {
-        return Ok(declared.to_path_buf());
-    }
-    let refusal = || {
-        format!(
-            "seed path '{}' resolves outside the working directory ({}); grant it with \
-             [read_paths] in the blueprint and your config, or move it inside",
-            declared.display(),
-            base.display()
-        )
-    };
-    if !read_paths.is_active() {
-        return Err(refusal());
-    }
-    // One arm for both ways this fails, because they fail the same way: a path
-    // that cannot be canonicalized is never matched, exactly as a canonical one
-    // the policy declines is not.
-    leviath_core::canonicalize_for_match(declared)
-        .filter(|c| {
-            matches!(
-                read_paths.decide(c),
-                leviath_core::ReadPathDecision::Allowed
-            )
-        })
-        .ok_or_else(refusal)
-}
-
-/// Resolve every region's initial content from its blueprint-declared
-/// [`RegionSeed`] plus the caller-provided values on `args`, into a
-/// name→content map ready for [`spawn_agent_seeded`].
-///
-/// The caller map is `{ "task": args.task } ∪ args.regions` (a `regions["task"]`
-/// wins). Then:
-/// - `CallerInput { name }` pulls from the caller map; if the region is
-///   `required` and the value is missing/blank this returns `Err` - the
-///   required-at-spawn gate, before any inference.
-/// - `Files` / `Glob` read workdir files; `Literal` is verbatim; `Rhai` runs a
-///   workdir script whose `String` return seeds the region.
-/// - `Command` runs a shell command in the workdir under `commands` -
-///   sandboxed, time- and size-capped, and skippable. Every failure is
-///   non-fatal unless the region is `required`.
-/// - Any caller key (other than `task`) that isn't a declared `CallerInput`
-///   region is rejected (typo protection, mirrors the CLI-side check).
-fn resolve_seeds(
-    blueprint: &Blueprint,
-    args: &SpawnArgs,
-    workdir: &str,
-    commands: &SeedCommandPolicy,
-    read_paths: &leviath_core::ReadPathPolicy,
-) -> Result<HashMap<String, String>, String> {
-    use leviath_core::layout::RegionSeed;
-
-    // The effective caller-supplied values: task text plus any named regions.
-    let mut caller: HashMap<String, String> = HashMap::new();
-    caller.insert("task".to_string(), args.task.clone());
-    for (k, v) in &args.regions {
-        caller.insert(k.clone(), v.clone());
-    }
-
-    // Unknown caller keys are tolerated here (silently unused): the CLI already
-    // rejects typos client-side in `resolve_spawn_args`, and an ACP host sending
-    // a stray `---region:...---` marker shouldn't fail the whole turn over it.
-
-    let base = std::path::Path::new(workdir);
-    let mut seeds: HashMap<String, String> = HashMap::new();
-
-    for region in &blueprint.context_layout.regions {
-        let Some(seed) = &region.seed else { continue };
-        match seed {
-            RegionSeed::CallerInput { name } => {
-                let value = caller.get(name).map(|s| s.as_str()).unwrap_or("");
-                if value.trim().is_empty() {
-                    if region.required {
-                        return Err(region.required_message.clone().unwrap_or_else(|| {
-                            format!(
-                                "required region '{}' was not provided; supply it via \
-                                 --{name} <text|@file> (CLI), a ---region:{name}--- block \
-                                 (ACP), or the API `regions` field",
-                                region.name
-                            )
-                        }));
-                    }
-                    // Optional and unprovided - leave the region empty.
-                    continue;
-                }
-                seeds.insert(region.name.clone(), value.to_string());
-            }
-            RegionSeed::Literal { text } => {
-                seeds.insert(region.name.clone(), text.clone());
-            }
-            RegionSeed::Files { paths } => {
-                let resolved = paths
-                    .iter()
-                    .map(|p| seed_path_within(base, &base.join(p), read_paths))
-                    .collect::<Result<Vec<_>, _>>()
-                    .map_err(|e| format!("region '{}': {e}", region.name))?;
-                let content = read_and_concat(&region.name, resolved.into_iter(), region.required)?;
-                if let Some(content) = content {
-                    seeds.insert(region.name.clone(), content);
-                }
-            }
-            RegionSeed::Glob { pattern } => {
-                let full = base.join(pattern);
-                let full = full.to_string_lossy();
-                let matches = glob::glob(&full)
-                    .map_err(|e| format!("region '{}': bad glob '{pattern}': {e}", region.name))?;
-                // Each *match* is checked, not the pattern: `../../*.toml`
-                // cannot be judged before it is expanded.
-                let paths = matches
-                    .filter_map(|m| m.ok())
-                    .map(|p| seed_path_within(base, &p, read_paths))
-                    .collect::<Result<Vec<_>, _>>()
-                    .map_err(|e| format!("region '{}': {e}", region.name))?;
-                let content = read_and_concat(&region.name, paths.into_iter(), region.required)?;
-                match content {
-                    Some(content) => {
-                        seeds.insert(region.name.clone(), content);
-                    }
-                    None if region.required => {
-                        return Err(format!(
-                            "required region '{}': glob '{pattern}' matched no files",
-                            region.name
-                        ));
-                    }
-                    None => {}
-                }
-            }
-            RegionSeed::Rhai { script } => {
-                let path = seed_path_within(base, &base.join(script), read_paths)
-                    .map_err(|e| format!("region '{}': {e}", region.name))?;
-                let src = std::fs::read_to_string(&path).map_err(|e| {
-                    format!(
-                        "region '{}': read rhai seed '{}': {e}",
-                        region.name,
-                        path.display()
-                    )
-                })?;
-                let mut input = rhai::Map::new();
-                input.insert("task".into(), rhai::Dynamic::from(args.task.clone()));
-                input.insert("workdir".into(), rhai::Dynamic::from(workdir.to_string()));
-                let out = leviath_scripting::ScriptEngine::new()
-                    .transform(&src, input)
-                    .map_err(|e| format!("region '{}': rhai seed failed: {e}", region.name))?;
-                if !out.trim().is_empty() {
-                    seeds.insert(region.name.clone(), out);
-                } else if region.required {
-                    return Err(format!(
-                        "required region '{}': rhai seed '{script}' returned empty",
-                        region.name
-                    ));
-                }
-            }
-            // A command seed *executes* at spawn, before any inference and so
-            // before any tool-approval prompt. It is therefore skipped outright
-            // when disabled, and every failure mode is non-fatal unless the
-            // region is `required` (mirroring the Files/Glob arms above): a
-            // discovery nicety must never be able to sink a run.
-            RegionSeed::Command { command } => {
-                if !commands.allowed {
-                    if region.required {
-                        return Err(format!(
-                            "required region '{}': command seeds are disabled \
-                             (`[security] allow_seed_commands = false` or --no-seed-commands)",
-                            region.name
-                        ));
-                    }
-                    tracing::warn!(
-                        region = %region.name,
-                        "command seed skipped: command seeds are disabled"
-                    );
-                    continue;
-                }
-                match commands.run(command, base) {
-                    Ok(out) if !out.trim().is_empty() => {
-                        seeds.insert(region.name.clone(), out);
-                    }
-                    Ok(_) => {
-                        if region.required {
-                            return Err(format!(
-                                "required region '{}': command seed '{command}' returned empty",
-                                region.name
-                            ));
-                        }
-                        tracing::warn!(
-                            region = %region.name,
-                            command = %command,
-                            "command seed returned no output; region left empty"
-                        );
-                    }
-                    Err(e) => {
-                        if region.required {
-                            return Err(format!(
-                                "required region '{}': command seed '{command}' failed: {e}",
-                                region.name
-                            ));
-                        }
-                        tracing::warn!(
-                            region = %region.name,
-                            command = %command,
-                            error = %e,
-                            "command seed failed; region left empty"
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(seeds)
-}
-
-/// Read each file and concatenate with `--- <path> ---` headers. Returns
-/// `Ok(None)` when the list is empty; a missing/unreadable file is an error only
-/// when `required`, else it is skipped.
-fn read_and_concat(
-    region: &str,
-    paths: impl Iterator<Item = std::path::PathBuf>,
-    required: bool,
-) -> Result<Option<String>, String> {
-    let mut parts: Vec<String> = Vec::new();
-    for path in paths {
-        match std::fs::read_to_string(&path) {
-            Ok(text) => parts.push(format!("--- {} ---\n{}", path.display(), text)),
-            Err(e) => {
-                if required {
-                    return Err(format!(
-                        "region '{region}': read seed file '{}': {e}",
-                        path.display()
-                    ));
-                }
-            }
-        }
-    }
-    Ok((!parts.is_empty()).then(|| parts.join("\n\n")))
-}
-
-/// Resolve an agent's `[read_paths]` declarations against the user's config
-/// into the policy its file tools enforce, plus a warning to surface when the
-/// declarations exist but nothing grants them.
-///
-/// A declared-but-ungranted agent still spawns - its out-of-workdir reads are
-/// refused per path with the same guidance - but the warning fires once here
-/// so the user learns about it at spawn rather than from a mid-run tool error.
-/// A malformed entry (in the blueprint or in the user's own grant list) is a
-/// hard spawn error: silently dropping it would either under-grant or run the
-/// agent with less vision than its author designed for.
-fn build_read_path_policy(
-    blueprint: &leviath_core::Blueprint,
-    config: &crate::config::Config,
-    workdir: &std::path::Path,
-) -> Result<(leviath_core::ReadPathPolicy, Option<String>), String> {
-    let Some(rp) = blueprint
-        .read_paths
-        .as_ref()
-        .filter(|rp| !rp.allow.is_empty())
-    else {
-        return Ok((leviath_core::ReadPathPolicy::inactive(), None));
-    };
-    let home = leviath_core::home_dir();
-    let declared =
-        leviath_core::ReadPathSet::compile(&rp.allow, workdir, home.as_deref(), cfg!(windows))
-            .map_err(|e| format!("agent '{}' [read_paths]: {e}", blueprint.name))?;
-    let grant_entries = config.read_path_grants_for_agent(&blueprint.name);
-    let grants =
-        leviath_core::ReadPathSet::compile(&grant_entries, workdir, home.as_deref(), cfg!(windows))
-            .map_err(|e| format!("read_paths grant in your config.toml: {e}"))?;
-    let allow_blueprint = config.security.allow_blueprint_read_paths;
-    let warning = (!allow_blueprint && grants.is_empty()).then(|| {
-        let entries = rp
-            .allow
-            .iter()
-            .map(|e| format!("\"{e}\""))
-            .collect::<Vec<_>>()
-            .join(", ");
-        format!(
-            "agent '{name}' declares [read_paths] but nothing grants them; reads outside \
-             the workdir will be refused. To grant them, add to your config.toml either:\n\
-             [security]\nallow_blueprint_read_paths = true\n\
-             or the specific paths:\n[agent_read_paths.{name}]\nallow = [{entries}]",
-            name = blueprint.name,
-        )
-    });
-    Ok((
-        leviath_core::ReadPathPolicy {
-            agent: blueprint.name.clone(),
-            blueprint: declared,
-            grants,
-            allow_blueprint,
-        },
-        warning,
-    ))
-}
-
-/// How many of a blueprint's `[read_paths]` entries the config grants, for the
-/// run listing. `None` when the blueprint declares none, and when the user's
-/// own grant list will not compile - that is a hard spawn error a line above,
-/// so there is no half-answer to record.
-fn read_path_grant_counts(
-    blueprint: &leviath_core::Blueprint,
-    config: &crate::config::Config,
-    workdir: &std::path::Path,
-) -> Option<leviath_core::run_meta::ReadPathGrantCounts> {
-    let report = crate::read_path_report::build(blueprint, config, workdir)?.ok()?;
-    Some(leviath_core::run_meta::ReadPathGrantCounts {
-        declared: report.declared(),
-        granted: report.granted(),
-    })
-}
-
-/// Raise the read tools to `Private` for an agent whose `[read_paths]` are
-/// actually granted: they can pull in content from outside the workdir -
-/// design docs, run archives, whatever else was granted - which the default
-/// `Internal` classification (written for workdir files) understates.
-fn bump_read_sensitivities(
-    map: &mut HashMap<String, leviath_core::TaintLevel>,
-    read_paths_granted: bool,
-) {
-    if !read_paths_granted {
-        return;
-    }
-    for tool in ["read_file", "read_files", "list_dir"] {
-        if let Some(level) = map.get_mut(tool) {
-            *level = (*level).max(leviath_core::TaintLevel::Private);
-        }
-    }
-}
+// Sections of the former single-file spawn path, one per question it answers.
+// The first two are re-exported because the daemon reaches them directly
+// (`model_defaults`, the script resolvers); the last two are internal to the
+// spawn path and only `build_agent_inner` calls them.
+mod policy;
+pub(crate) use policy::*;
+mod scripts;
+pub(crate) use scripts::*;
+mod seeds;
+use seeds::*;
+mod tool_state;
+use tool_state::*;
 
 /// Everything a spawn needs that is not the request itself.
 ///
